Removing publishConfig
The publishConfig specified the public npm registry. This is the default registry and is also specified in the npm publish github action, so publishConfig within the package is not necessary. Additionally, publishConfig overrides the target registry specified in any other way (.npmrc, cli, github action), making it impossible to publish to test or private registries, without manually modifying the package.